+++ CHANGELOG.md
@@ -17,6 +17,7 @@ CHANGELOG
 * Updated projet to use npm version of videojs-swf ([view](https://github.com/videojs/video.js/pull/930))
 * Added support for dist zipping on windows ([view](https://github.com/videojs/video.js/pull/944))
 * Fixed iOS fullscreen issue ([view](https://github.com/videojs/video.js/pull/977))
+* Fixed touch event bubbling ([view](https://github.com/videojs/video.js/pull/992))
 
 --------------------